requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
colorama==0.4.4
decorator==5.1.1
Django==3.2.12
django-bootstrap-v5==1.0.11
django-bootstrap-pagination==1.7.1
django-dotenv==1.4.2
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
pure-eval==0.2.2
Pygments==2.11.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
tzdata==2021.5
wcwidth==0.2.5
